server1/requirements.txt
flask
requests
pillow
opencv-python
numpy
pillow-heif
boto3
